+++ Reflection/src/utopia/reflection/component/drawing/template/ImageDrawerLike.scala
@@ -46,31 +46,18 @@ trait ImageDrawerLike extends CustomDrawer
 	{
 		// Calculates the size of the drawn image
 		val defaultSize = image.size + insets.optimal.total
-		val imageToDraw =
-		{
+		val imageToDraw = {
 			// Default case: No upscaling used or required
-			if (!useUpscaling || defaultSize.height >= bounds.height || defaultSize.width >= bounds.width)
-			{
-				// Checks whether downscaling is required
-				val minNaturalSize = image.size + insets.min.total
-				if (minNaturalSize.height > bounds.height || minNaturalSize.width > bounds.width)
-				{
-					val availableSize = bounds.size - insets.min.total
-					if (availableSize.isPositive)
-						Some(image.fitting(bounds.size - insets.min.total))
-					else
-						None
-				}
-				else
-					Some(image)
+			if (!useUpscaling || defaultSize.existsDimensionWith(bounds.size) { _ >= _ }) {
+				// Downscales if necessary
+				(bounds.size - insets.min.total).ifPositive.map { image.fittingWithin(_) }
 			}
-			else
-			{
-				// Case: Upscaling is required (still limited by original image resolution, unless original image is
-				// already over source resolution)
+			else {
+				// Case: Upscaling is required (still limited by original image resolution,
+				// unless original image is already over source resolution)
 				val imageSize = (bounds.size - insets.mapToInsets { l => l.max.getOrElse(l.optimal) }.total) min
 					(image.size max image.sourceResolution)
-				Some(image.fitting(imageSize))
+				Some(image.fittingWithin(imageSize, maximize = true))
 			}
 		}